make Radisk available in browser for @fuchidahiro
