+++ lib/load.js
@@ -1,9 +1,9 @@
 if(typeof window !== "undefined"){
   var Gun = window.Gun;
 } else {
-  var Gun = require('gun/gun');
+  var Gun = require('../gun');
 }
-Gun.chain.open || require('gun/lib/open');
+Gun.chain.open || require('./open');
 
 Gun.chain.load = function(cb, opt, at){
 	(opt = opt || {}).off = !0;
