+++ lib/open.js
@@ -1,7 +1,7 @@
 if(typeof window !== "undefined"){
   var Gun = window.Gun;
 } else { 
-  var Gun = require('gun/gun');
+  var Gun = require('../gun');
 }
 
 Gun.chain.open = function(cb, opt, at){
