+++ lib/radix.js
@@ -1,8 +1,4 @@
-var Gun = require('../gun');
 ;(function(){
-	var map = Gun.obj.map, no = {}, u;
-
-	var $ = String.fromCharCode(30), _ = String.fromCharCode(29);
 
 	function Radix(){
 		var radix = function(key, val, t){
@@ -46,6 +42,7 @@ var Gun = require('../gun');
 		}
 		return radix;
 	};
+
 	Radix.map = function map(radix, cb, opt, pre){ pre = pre || []; 
 		var t = radix[_] || radix, keys = Object.keys(t).sort(), i = 0, l = keys.length;
 		for(;i < l; i++){ var key = keys[i], tree = t[key], tmp;
@@ -64,7 +61,18 @@ var Gun = require('../gun');
 			}
 		}
 	};
+
 	Object.keys = Object.keys || function(o){ return map(o, function(v,k,t){t(k)}) }
+
+	if(typeof window !== "undefined"){
+	  var Gun = window.Gun;
+	  window.Radix = Radix;
+	} else { 
+	  var Gun = require('../gun');
+		try{ module.exports = Radix }catch(e){}
+	}
 	
-	module.exports = Radix;
+	var map = Gun.obj.map, no = {}, u;
+	var $ = String.fromCharCode(30), _ = String.fromCharCode(29);
+	
 }());
