+++ lib/store.js
@@ -1,17 +1,21 @@
-var Gun = require('../gun');
-var Radisk = require('./radisk');
-var fs = require('fs');
-var Radix = Radisk.Radix;
-var u;
+if(typeof window === "undefined"){
+  var Gun = require('../gun');
+}
 
 Gun.on('opt', function(ctx){
 	this.to.next(ctx);
-	var opt = ctx.opt;
+	var opt = ctx.opt, u;
+	if(typeof window !== "undefined"){
+		opt.window = window;
+	}
 	if(ctx.once){ return }
-	if(false !== opt.localStorage && !process.env.AWS_S3_BUCKET){ return } // TODO: Remove this after migration.
+	if(false !== opt.localStorage && !(!opt.window && process.env.AWS_S3_BUCKET)){ return } // TODO: Remove this after migration.
 	if(false === opt.radisk){ return }
 	console.log("BUG WARNING: Radix Storage Engine (RAD) has a known rare edge case, if data gets split between file chunks, a GET may only return the first chunk!!!");
-	opt.store = opt.store || Store(opt);
+	var Radisk = (opt.window && opt.window.Radisk) || require('./radisk');
+	var Radix = Radisk.Radix;
+
+	opt.store = opt.store || (!opt.window && require('./rfs')(opt));
 	var rad = Radisk(opt);
 
 	ctx.on('put', function(at){
@@ -53,80 +57,3 @@ Gun.on('opt', function(ctx){
 	});
 
 });
-
-function Store(opt){
-	opt = opt || {};
-	opt.file = String(opt.file || 'radata');
-
-	var store = function Store(){};
-	store.put = function(file, data, cb){
-		var random = Math.random().toString(36).slice(-3)
-		fs.writeFile(opt.file+'-'+random+'.tmp', data, function(err, ok){
-			if(err){ return cb(err) }
-			move(opt.file+'-'+random+'.tmp', opt.file+'/'+file, cb);
-		});
-	};
-	store.get = function(file, cb){
-		fs.readFile(opt.file+'/'+file, function(err, data){
-			if(err){
-				if('ENOENT' === (err.code||'').toUpperCase()){
-					return cb(null);
-				}
-				Gun.log("ERROR:", err)
-			}
-			if(data){ data = data.toString() }
-			cb(err, data);
-		});
-	};
-	store.list = function(cb, match){
-		fs.readdir(opt.file, function(err, dir){
-			Gun.obj.map(dir, cb) || cb(); // Stream interface requires a final call to know when to be done.
-		});
-	};
-	if(!fs.existsSync(opt.file)){ fs.mkdirSync(opt.file) }
-	//store.list(function(){ return true });
-	return store;
-}
-
-function move(oldPath, newPath, cb) {
-	fs.rename(oldPath, newPath, function (err) {
-		if (err) {
-			if (err.code === 'EXDEV') {
-				var readStream = fs.createReadStream(oldPath);
-				var writeStream = fs.createWriteStream(newPath);
-
-				readStream.on('error', cb);
-				writeStream.on('error', cb);
-
-				readStream.on('close', function () {
-					fs.unlink(oldPath, cb);
-				});
-
-				readStream.pipe(writeStream);
-			} else {
-				cb(err);
-			}
-		} else {
-			cb();
-		}
-	});
-};
-
-module.exports = Store;
-
-
-;(function(){
-	return;
-	process.env.AWS_S3_BUCKET = 'test-s3';
-	process.env.AWS_ACCESS_KEY_ID = 'asdf';
-	process.env.AWS_SECRET_ACCESS_KEY = 'fdsa';
-	process.env.fakes3 = 'http://localhost:4567';
-	process.env.AWS_S3_THROTTLE = 0;
-
-	return;
-	global.Gun = require('../gun');
-	//require('./rs3');
-	
-	
-	require('../test/abc');
-}());
