+++ lib/time.js
@@ -1,5 +1,5 @@
 if(typeof window === "undefined"){ //Not in the browser, Include from node
-  var Gun = require('gun/gun');
+  var Gun = require('../gun');
 }
 
 ;(function(){
