+++ lib/upload.js
@@ -27,7 +27,11 @@
 		var file = (((e.event || e).target || e).result || e), img = new Image();
     img.src = file;
     img.onload = function(){
-      if(!h && img.width > w){ h = img.height * (w / img.width) }
+    	if(img.width < w && img.height < (h||Infinity)){
+    		e.base64 = file;
+    		return cb(e || file);
+    	}
+      if(!h){ h = img.height * (w / img.width) }
 			var canvas = document.createElement('canvas'), ctx = canvas.getContext('2d');
 	    canvas.width = w;
 	    canvas.height = h;
