+++ test/panic/radisk.js
@@ -3,10 +3,12 @@ var config = {
 	port: 8080,
 	servers: 2,
 	browsers: 2,
-	each: 2000,
+	each: 100,
 	burst: 50,
 	wait: 1,
 	dir: __dirname,
+	chunk: 1024 * 10,
+	notrad: false,
 	route: {
 		'/': __dirname + '/index.html',
 		'/gun.js': __dirname + '/../../gun.js',
@@ -50,25 +52,19 @@ describe("Make sure the Radix Storage Engine (RSE) works.", function(){
 	it("GUN started!", function(){
 		return server.run(function(test){
 			var env = test.props;
-			console.log("????", process.argv);
 			test.async();
 			if(require('fs').existsSync('radata')){
 				console.log("Please delete previous data first!");
 				explode;
 				return;
 			}
-			/*setInterval(function(){
-				var mem = process.memoryUsage();
-				var u = Math.round(mem.heapUsed / 1024 / 1024 * 100) / 100;
-				console.log(u, 'MB of', Math.round(mem.heapTotal / 1024 / 1024 * 100) / 100);
-			}, 1000);*/
 			var port = env.config.port + env.i;
 			var server = require('http').createServer(function(req, res){
 				res.end("I am "+ env.i +"!");
 			});
 			var Gun = require('gun');
 			//require('gun/lib/store');
-			var gun = Gun({web: server, localStorage: false, until: 1, memory: 50, chunk: 1024 * 100});
+			var gun = Gun({web: server, localStorage: env.config.notrad, until: 1, memory: 50, chunk: env.config.chunk, file: 'radata'});
 			server.listen(port, function(){
 				test.done();
 			});
@@ -86,7 +82,7 @@ describe("Make sure the Radix Storage Engine (RSE) works.", function(){
 			console.log("I AM ALICE");
 			localStorage.clear();
 			var env = test.props;
-			var gun = Gun({peers: 'http://'+ env.config.IP + ':' + (env.config.port + 1) + '/gun', localStorage: false});
+			var gun = Gun({peers: 'http://'+ env.config.IP + ':' + (env.config.port + 1) + '/gun', localStorage: env.config.notrad});
 			window.gun = gun;
 
 			var n = Gun.time.is(), i = 0, c = 0, b = env.config.burst, l = env.config.each;
@@ -154,7 +150,7 @@ describe("Make sure the Radix Storage Engine (RSE) works.", function(){
 			});
 			var Gun = require('gun');
 			//require('gun/lib/store');
-			var gun = Gun({web: server, localStorage: false, until: 1, memory: 50, chunk: 1024 * 100});
+			var gun = Gun({web: server, localStorage: env.config.notrad, until: 1, memory: 50, chunk: env.config.notrad, file: 'radata'});
 			server.listen(port, function(){
 				test.done();
 			});
@@ -167,11 +163,12 @@ describe("Make sure the Radix Storage Engine (RSE) works.", function(){
 			console.log("I AM BOB");
 			localStorage.clear();
 			var env = test.props;
-			var gun = Gun({peers: 'http://'+ env.config.IP + ':' + (env.config.port + 2) + '/gun', localStorage: false});
+			var gun = Gun({peers: 'http://'+ env.config.IP + ':' + (env.config.port + 2) + '/gun', localStorage: env.config.notrad});
 			window.gun = gun;
 
 			var n = Gun.time.is(), i = 0, c = 0, b = env.config.burst, l = env.config.each;
 			var raw = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
+			window.FOO = [];
 
 			function check(i){
 				if(i > l){
@@ -183,7 +180,8 @@ describe("Make sure the Radix Storage Engine (RSE) works.", function(){
 					if((raw+i) !== data.hello){ return test.fail('wrong ' + i) }
 					if(d){ return } d = true;
 					//!(c % b) && 
-					console.log(c+'/'+l);//, '@'+Math.floor(b/((-n + (n = Gun.time.is()))/1000))+'/sec'));
+					window.FOO.push(i);
+					console.log(c+'/'+l, 'yeah?', i, Gun.node.soul(data));//, '@'+Math.floor(b/((-n + (n = Gun.time.is()))/1000))+'/sec'));
 					window.GOT = c++;
 					//localStorage.clear();
 					ref.off();
